// services/execd/tests/generated/v1/configd.ts
export interface ConsumerBinding {
  placement: PlacementBinding | undefined;
  consumerId: string;
  purpose: string;
}
// ...
export interface PublishConfigurationResponse {
  configuration: ConfigurationMetadata | undefined;
  version: ConfigurationVersionMetadata | undefined;
}

export interface ResolveConfigurationRequest {
  configurationId: string;
  configurationVersionId: string;
  binding: ConsumerBinding | undefined;
}

export interface ResolveConfigurationResponse {
  configuration: ConfigurationMetadata | undefined;
  version: ConfigurationVersionMetadata | undefined;
  contentJson: Buffer;
}

export interface ConfigurationMetadata {
  configurationId: string;
  binding: ConsumerBinding | undefined;
  currentConfigurationVersionId: string;
  revision: bigint;
  createdAt: Date | undefined;
  updatedAt: Date | undefined;
}

export interface ConfigurationVersionMetadata {
  configurationVersionId: string;
  configurationId: string;
  contentLength: number;
  contentSha256: Buffer;
  createdAt: Date | undefined;
}
// ...
export const ConsumerBinding: MessageFns<ConsumerBinding> = {
  encode(message: ConsumerBinding, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.placement !== undefined) {
      PlacementBinding.encode(message.placement, writer.uint32(10).fork()).join();
    }
    if (message.consumerId !== "") {
      writer.uint32(18).string(message.consumerId);
    }
    if (message.purpose !== "") {
      writer.uint32(26).string(message.purpose);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ConsumerBinding {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseConsumerBinding();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.placement = PlacementBinding.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.consumerId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.purpose = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ConsumerBinding {
    return {
      placement: isSet(object.placement) ? PlacementBinding.fromJSON(object.placement) : undefined,
      consumerId: isSet(object.consumerId) ? globalThis.String(object.consumerId) : "",
      purpose: isSet(object.purpose) ? globalThis.String(object.purpose) : "",
    };
  },

  toJSON(message: ConsumerBinding): unknown {
    const obj: any = {};
    if (message.placement !== undefined) {
      obj.placement = PlacementBinding.toJSON(message.placement);
    }
    if (message.consumerId !== "") {
      obj.consumerId = message.consumerId;
    }
    if (message.purpose !== "") {
      obj.purpose = message.purpose;
    }
    return obj;
  },

  create(base?: DeepPartial<ConsumerBinding>): ConsumerBinding {
    return ConsumerBinding.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<ConsumerBinding>): ConsumerBinding {
    const message = createBaseConsumerBinding();
    message.placement = (object.placement !== undefined && object.placement !== null)
      ? PlacementBinding.fromPartial(object.placement)
      : undefined;
    message.consumerId = object.consumerId ?? "";
    message.purpose = object.purpose ?? "";
    return message;
  },
};
// ...
export const PublishConfigurationResponse: MessageFns<PublishConfigurationResponse> = {
  encode(message: PublishConfigurationResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.configuration !== undefined) {
      ConfigurationMetadata.encode(message.configuration, writer.uint32(10).fork()).join();
    }
    if (message.version !== undefined) {
      ConfigurationVersionMetadata.encode(message.version, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): PublishConfigurationResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePublishConfigurationResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.configuration = ConfigurationMetadata.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.version = ConfigurationVersionMetadata.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PublishConfigurationResponse {
    return {
      configuration: isSet(object.configuration) ? ConfigurationMetadata.fromJSON(object.configuration) : undefined,
      version: isSet(object.version) ? ConfigurationVersionMetadata.fromJSON(object.version) : undefined,
    };
  },

  toJSON(message: PublishConfigurationResponse): unknown {
    const obj: any = {};
    if (message.configuration !== undefined) {
      obj.configuration = ConfigurationMetadata.toJSON(message.configuration);
    }
    if (message.version !== undefined) {
      obj.version = ConfigurationVersionMetadata.toJSON(message.version);
    }
    return obj;
  },

  create(base?: DeepPartial<PublishConfigurationResponse>): PublishConfigurationResponse {
    return PublishConfigurationResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<PublishConfigurationResponse>): PublishConfigurationResponse {
    const message = createBasePublishConfigurationResponse();
    message.configuration = (object.configuration !== undefined && object.configuration !== null)
      ? ConfigurationMetadata.fromPartial(object.configuration)
      : undefined;
    message.version = (object.version !== undefined && object.version !== null)
      ? ConfigurationVersionMetadata.fromPartial(object.version)
      : undefined;
    return message;
  },
};
// ...
export const ConfigurationMetadata: MessageFns<ConfigurationMetadata> = {
  encode(message: ConfigurationMetadata, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.configurationId !== "") {
      writer.uint32(10).string(message.configurationId);
    }
    if (message.binding !== undefined) {
      ConsumerBinding.encode(message.binding, writer.uint32(18).fork()).join();
    }
    if (message.currentConfigurationVersionId !== "") {
      writer.uint32(26).string(message.currentConfigurationVersionId);
    }
    if (message.revision !== 0n) {
      if (BigInt.asUintN(64, message.revision) !== message.revision) {
        throw new globalThis.Error("value provided for field message.revision of type uint64 too large");
      }
      writer.uint32(32).uint64(message.revision);
    }
    if (message.createdAt !== undefined) {
      Timestamp.encode(toTimestamp(message.createdAt), writer.uint32(42).fork()).join();
    }
    if (message.updatedAt !== undefined) {
      Timestamp.encode(toTimestamp(message.updatedAt), writer.uint32(50).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ConfigurationMetadata {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseConfigurationMetadata();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.configurationId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.binding = ConsumerBinding.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.currentConfigurationVersionId = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.revision = reader.uint64() as bigint;
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.createdAt = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.updatedAt = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ConfigurationMetadata {
    return {
      configurationId: isSet(object.configurationId) ? globalThis.String(object.configurationId) : "",
      binding: isSet(object.binding) ? ConsumerBinding.fromJSON(object.binding) : undefined,
      currentConfigurationVersionId: isSet(object.currentConfigurationVersionId)
        ? globalThis.String(object.currentConfigurationVersionId)
        : "",
      revision: isSet(object.revision) ? BigInt(object.revision) : 0n,
      createdAt: isSet(object.createdAt) ? fromJsonTimestamp(object.createdAt) : undefined,
      updatedAt: isSet(object.updatedAt) ? fromJsonTimestamp(object.updatedAt) : undefined,
    };
  },

  toJSON(message: ConfigurationMetadata): unknown {
    const obj: any = {};
    if (message.configurationId !== "") {
      obj.configurationId = message.configurationId;
    }
    if (message.binding !== undefined) {
      obj.binding = ConsumerBinding.toJSON(message.binding);
    }
    if (message.currentConfigurationVersionId !== "") {
      obj.currentConfigurationVersionId = message.currentConfigurationVersionId;
    }
    if (message.revision !== 0n) {
      obj.revision = message.revision.toString();
    }
    if (message.createdAt !== undefined) {
      obj.createdAt = message.createdAt.toISOString();
    }
    if (message.updatedAt !== undefined) {
      obj.updatedAt = message.updatedAt.toISOString();
    }
    return obj;
  },

  create(base?: DeepPartial<ConfigurationMetadata>): ConfigurationMetadata {
    return ConfigurationMetadata.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<ConfigurationMetadata>): ConfigurationMetadata {
    const message = createBaseConfigurationMetadata();
    message.configurationId = object.configurationId ?? "";
    message.binding = (object.binding !== undefined && object.binding !== null)
      ? ConsumerBinding.fromPartial(object.binding)
      : undefined;
    message.currentConfigurationVersionId = object.currentConfigurationVersionId ?? "";
    message.revision = object.revision ?? 0n;
    message.createdAt = object.createdAt ?? undefined;
    message.updatedAt = object.updatedAt ?? undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | bigint | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}